lib: separate data definition and parsing code
The `Security.Advisories` module is a bit unweildy and is going to
get worse. Extract data definitions and parsing to separate modules.
