+++ code/hsec-tools/hsec-tools.cabal
@@ -27,9 +27,10 @@ tested-with:
   GHC ==8.10.7 || ==9.0.2 || ==9.2.7 || ==9.4.5 || ==9.6.2
 
 library
-    exposed-modules:
-                      Security.Advisories,
-                      Security.OSV
+    exposed-modules:    Security.Advisories
+                      , Security.Advisories.Definition
+                      , Security.Advisories.Parse
+                      , Security.OSV
     build-depends:    base >=4.14 && < 4.19,
                       text >= 1.2 && < 3,
                       time >= 1.9 && < 1.14,
