+++ code/hsec-tools/src/Security/Advisories/Definition.hs
@@ -1,0 +1,93 @@
+{-# LANGUAGE DerivingVia #-}
+
+module Security.Advisories.Definition
+  ( Advisory(..)
+    -- * Supporting types
+  , CWE(..)
+  , Architecture(..)
+  , AffectedVersionRange(..)
+  , OS(..)
+  , Date(..)
+  , Keyword(..)
+  )
+  where
+
+import Data.Text (Text)
+import Distribution.Types.VersionRange (VersionRange)
+
+import Text.Pandoc.Definition (Pandoc)
+
+import Security.OSV (Reference)
+
+data Advisory = Advisory
+  { advisoryId :: Text
+  , advisoryPackage :: Text
+  , advisoryDate :: Date
+  , advisoryCWEs :: [CWE]
+  , advisoryKeywords :: [Keyword]
+  , advisoryAliases :: [Text]
+  , advisoryCVSS :: Text
+  , advisoryVersions :: [AffectedVersionRange]
+  , advisoryArchitectures :: Maybe [Architecture]
+  , advisoryOS :: Maybe [OS]
+  , advisoryNames :: [(Text, VersionRange)]
+  , advisoryReferences :: [Reference]
+  , advisoryPandoc :: Pandoc  -- ^ Parsed document, without TOML front matter
+  , advisoryHtml :: Text
+  , advisorySummary :: Text
+  }
+  deriving stock (Show)
+
+newtype CWE = CWE {unCWE :: Integer}
+  deriving stock (Show)
+
+data Architecture
+  = AArch64
+  | Alpha
+  | Arm
+  | HPPA
+  | HPPA1_1
+  | I386
+  | IA64
+  | M68K
+  | MIPS
+  | MIPSEB
+  | MIPSEL
+  | NIOS2
+  | PowerPC
+  | PowerPC64
+  | PowerPC64LE
+  | RISCV32
+  | RISCV64
+  | RS6000
+  | S390
+  | S390X
+  | SH4
+  | SPARC
+  | SPARC64
+  | VAX
+  | X86_64
+  deriving stock (Show)
+
+data OS
+  = Windows
+  | MacOS
+  | Linux
+  | FreeBSD
+  | Android
+  | NetBSD
+  | OpenBSD
+  deriving stock (Show)
+
+data Date = Date {dateYear :: Integer, dateMonth :: Int, dateDay :: Int}
+  deriving stock (Show)
+
+newtype Keyword = Keyword Text
+  deriving stock (Eq, Ord)
+  deriving (Show) via Text
+
+data AffectedVersionRange = AffectedVersionRange
+  { affectedVersionRangeIntroduced :: Text,
+    affectedVersionRangeFixed :: Maybe Text
+  }
+  deriving stock (Show)
